requirements.txt
streamlit==1.38.0
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.1
